Inline followed effort ids instead of fetching them
The spread wrapper renders outside the fragment-cached table body, so
the presenter can supply the signed-in user's followed effort ids
directly as a Stimulus array value. This removes the JSON endpoint,
its route and policy, and the client-side fetch, and saves a round
trip per signed-in spread load.
